coming along
basic view models and service created and some more work done on the ui
so you can add a new host, but it does nothing apart from adding it to
the UI

+++ src/PingIt.Domain/PingHistory.cs
@@ -6,6 +6,6 @@ namespace PingIt.Domain
     {
         public DateTime PingedAt { get; set; }
         public PingStatus Status { get; set; }
-        public int Duration { get; set; }
+        public long Duration { get; set; }
     }
 }
